cleanup and split into crates
