+++ shielder/contract/errors.rs
@@ -1,4 +1,5 @@
 use crate::traits::psp22::PSP22Error;
+use mocked_zk::errors::ZkpError;
 
 #[ink::scale_derive(Encode, Decode, TypeInfo)]
 #[derive(PartialEq, Debug)]
@@ -16,5 +17,11 @@ pub enum ShielderError {
 impl From<PSP22Error> for ShielderError {
     fn from(inner: PSP22Error) -> Self {
         ShielderError::PSP22(inner)
+    }
+}
+
+impl From<ZkpError> for ShielderError {
+    fn from(_inner: ZkpError) -> Self {
+        ShielderError::ZkpVerificationFail
     }
 }
