+++ shielder/mocked_zk/Cargo.toml
@@ -1,24 +1,20 @@
 [package]
-name = "shielder-contract"
+name = "mocked_zk"
 version = "0.1.0"
 authors = ["Cardinal"]
 homepage = "https://alephzero.org"
 repository = "https://github.com/Cardinal-Cryptography/zk-apps"
 edition = "2021"
 
-[dependencies]
-# ink = { workspace = true }
-ink = { version = "5.0.0-rc", default-features = false }
-mocked_zk = { path = "../mocked_zk", default-features = false }
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[lib]
-path = "lib.rs"
+[dependencies]
+ink = { version = "5.0.0-rc", default-features = false }
+serde = { version = "1.0.197", default-features = false, features = ["derive"] }
 
 [features]
 default = ["std"]
 std = [
     "ink/std",
-    "mocked_zk/std",
 ]
 ink-as-dependency = []
-e2e-tests = []
