+++ shielder/mocked_zk/src/errors.rs
@@ -1,0 +1,7 @@
+#[ink::scale_derive(Encode, Decode, TypeInfo)]
+#[derive(PartialEq, Debug)]
+pub enum ZkpError {
+    AccountUpdateError,
+    OperationCombineError,
+    VerificationError,
+}
