+++ shielder/mocked_zk/src/relations.rs
@@ -1,14 +1,11 @@
-use super::{
+use crate::{
     account::Account,
+    combine_merkle_hash,
+    errors::ZkpError,
     note::Note,
     ops::{OpPriv, OpPub, Operation},
     traits::Hashable,
-    Scalar,
-};
-use crate::{
-    contract::{MERKLE_TREE_DEPTH, TOKENS_NUMBER},
-    errors::ShielderError,
-    merkle::{self},
+    Scalar, MERKLE_TREE_DEPTH, TOKENS_NUMBER,
 };
 
 /// mocked proof of knowledge, not ZK
@@ -28,10 +25,10 @@ pub struct ZkProof {
     merkle_proof_leaf_id: u32,
 }
 
-pub fn verify_hash<T: Hashable>(to_hash: T, hash: Scalar) -> Result<Scalar, ShielderError> {
+pub fn verify_hash<T: Hashable>(to_hash: T, hash: Scalar) -> Result<Scalar, ZkpError> {
     let real_hash = to_hash.hash();
     if real_hash != hash {
-        return Err(ShielderError::ZkpVerificationFail);
+        return Err(ZkpError::VerificationError);
     }
     Ok(real_hash)
 }
@@ -86,7 +83,7 @@ impl ZkProof {
         nullifier: Scalar,
         merkle_proof: [Scalar; MERKLE_TREE_DEPTH],
         merkle_proof_leaf_id: u32,
-    ) -> Result<(Scalar, Self), ShielderError> {
+    ) -> Result<(Scalar, Self), ZkpError> {
         let acc_updated = self.acc_new.update(operation)?;
         let note = Note::new(self.id, trapdoor, nullifier, acc_updated.hash());
         let new_proof = self.transition(
@@ -104,29 +101,25 @@ impl ZkProof {
         &self,
         op: Operation,
         h_acc_old: Scalar,
-    ) -> Result<Account, ShielderError> {
+    ) -> Result<Account, ZkpError> {
         let acc_new = self.acc_old.update(op)?;
         verify_hash(self.acc_old, h_acc_old)?;
         Ok(acc_new)
     }
 
-    fn verify_merkle_proof(
-        &self,
-        h_note_old: Scalar,
-        merkle_root: Scalar,
-    ) -> Result<(), ShielderError> {
+    fn verify_merkle_proof(&self, h_note_old: Scalar, merkle_root: Scalar) -> Result<(), ZkpError> {
         let mut id = self.merkle_proof_leaf_id;
         let mut scalar = h_note_old;
         for node in self.merkle_proof {
             if id % 2 == 0 {
-                scalar = merkle::compute_hash(scalar, node);
+                scalar = combine_merkle_hash(scalar, node);
             } else {
-                scalar = merkle::compute_hash(node, scalar);
+                scalar = combine_merkle_hash(node, scalar);
             }
             id /= 2;
         }
         if scalar != merkle_root {
-            return Err(ShielderError::ZkpVerificationFail);
+            return Err(ZkpError::VerificationError);
         }
         Ok(())
     }
@@ -135,7 +128,7 @@ impl ZkProof {
         &self,
         h_note_new: Scalar,
         tokens_list: [Scalar; TOKENS_NUMBER],
-    ) -> Result<(), ShielderError> {
+    ) -> Result<(), ZkpError> {
         let h_acc_new = Account::new(tokens_list).hash();
         let note_new = Note::new(self.id, self.trapdoor_new, self.nullifier_new, h_acc_new);
         verify_hash(note_new, h_note_new)?;
@@ -148,7 +141,7 @@ impl ZkProof {
         h_note_new: Scalar,
         merkle_root: Scalar,
         nullifier_old: Scalar,
-    ) -> Result<(), ShielderError> {
+    ) -> Result<(), ZkpError> {
         let h_acc_old = self.acc_old.hash();
         let op = Operation::combine(op_pub, self.op_priv)?;
         let acc_new = self.verify_acccount_update(op, h_acc_old)?;
