+++ shielder/mocked_zk/src/traits.rs
@@ -1,4 +1,4 @@
-use crate::types::Scalar;
+use crate::Scalar;
 
 pub trait Hashable {
     fn hash(&self) -> Scalar;
